+++ packages/amplify-gen2-codegen/src/index.ts
@@ -108,6 +108,7 @@ export const createGen2Renderer = ({
         module: 'es2022',
         moduleResolution: 'bundler',
         resolveJsonModule: true,
+        // eslint-disable-next-line spellcheck/spell-checker
         esModuleInterop: true,
         forceConsistentCasingInFileNames: true,
         strict: true,