+++ packages/amplify-gen2-codegen/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-next-6.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-gen2-codegen@0.1.0-next-5.0...@aws-amplify/amplify-gen2-codegen@0.1.0-next-6.0) (2025-03-21)
+
+
+### Bug Fixes
+
+* add idp codegen and fix a couple of minor bugs ([2b1d203](https://github.com/aws-amplify/amplify-cli/commit/2b1d203f2c065a701076cf4cdbd50b4ef51bc7ee))
+* **amplify-gen2-codegen:** remove extraneous newline ([aaf6dba](https://github.com/aws-amplify/amplify-cli/commit/aaf6dba696091933bc99583a2262f23dc96ea0ec))
+* migration qa feedback ([0f50d1c](https://github.com/aws-amplify/amplify-cli/commit/0f50d1c0ff2607c63cd3bcdd1589d38940a2b6d4))
+
+
+
+
+
 # [0.1.0-next-5.0](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-gen2-codegen@0.1.0-next-4.0...@aws-amplify/amplify-gen2-codegen@0.1.0-next-5.0) (2025-03-19)